refactor: day and logger
