+++ extensions/iceworks-time-master/src/managers/walkClock.ts
@@ -1,8 +1,19 @@
-import { setNowDay, checkIsNewDay } from '../utils/time';
+import { setNowDay, checkIsNewDay } from '../utils/day';
 import { sendPayload, checkPayloadIsLimited } from '../utils/sender';
 import { checkStorageDaysIsLimited } from '../utils/storage';
-import logger, { checkLogsIsLimited, reloadLogger } from '../utils/logger';
+import logger from '../utils/logger';
 import { checkMidnightDurationMins, sendPayloadDurationMins } from '../config';
+
+async function checkIsLimited() {
+  try {
+    await Promise.all([
+      checkStorageDaysIsLimited(),
+      checkPayloadIsLimited(),
+    ]);
+  } catch (e) {
+    logger.error('[walkClock][checkIsLimited] got error:', e);
+  }
+}
 
 export async function checkMidnight() {
   const isNewDay = checkIsNewDay();
@@ -10,16 +21,9 @@ export async function checkMidnight() {
 
   if (isNewDay) {
     setNowDay();
-    reloadLogger();
     try {
-      await Promise.all([
-        checkLogsIsLimited(),
-        checkStorageDaysIsLimited(),
-        (async function () {
-          await checkPayloadIsLimited();
-          await sendPayload(true);
-        })(),
-      ]);
+      await checkIsLimited();
+      await sendPayload();
     } catch (e) {
       logger.error('[walkClock][checkMidnight] got error:', e);
     }
@@ -51,5 +55,5 @@ export async function deactivate() {
   if (sendPayloadTimer) {
     clearInterval(sendPayloadTimer);
   }
-  await sendPayload(true);
+  await sendPayload();
 }
