+++ extensions/iceworks-time-master/src/storages/user.ts
@@ -3,7 +3,8 @@ import * as fse from 'fs-extra';
 import * as moment from 'moment';
 import { getStorageReportsPath, getStorageDayPath } from '../utils/storage';
 import { getRangeReport } from '../utils/report';
-import { getDay, getLastWeekDays } from '../utils/time';
+import { getLastWeekDays } from '../utils/time';
+import { getDay } from '../utils/day';
 import { updateAverageSummary } from './average';
 import { jsonSpaces } from '../config';
 import logger from '../utils/logger';
